project no 4 Hiring Process analytics
In this project we are going to extract a statistical insights from the data, which helps to find trends,patterns and styles of hiring process and departments and employees and gender distributinos and so on. understanding all these insights from the perspective of a analyst is essential to share the to company about future trendz and predictions.